dom/bindings/test/test_bug759621.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=759621
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 759621</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=759621">Mozilla Bug 759621</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 759621 **/
var l = document.getElementsByTagName("*");
l.namedItem = "pass";
is(l.namedItem, "pass", "Should be able to set expando shadowing a proto prop");

</script>
</pre>
</body>
</html>
